@@ -32,13 +32,16 @@
     "axios": "^0.19.2",
     "classnames": "^2.2.5",
     "date-fns": "^1.28.2",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "html-webpack-plugin": "^4.3.0",
     "lodash": "^4.17.4",
     "normalizr": "^3.1.0",
     "parse-link-header": "^1.0.1",
     "query-string": "^5.0.0",
     "react": "^16.0.0",
+    "react-addons-test-utils": "^15.6.2",
     "react-document-title": "^2.0.2",
-    "react-dom": "^16.0.0",
+    "react-dom": "^16.13.1",
     "react-redux": "^5.0.2",
     "react-router-dom": "^4.0.0",
     "redux": "^3.6.0",
@@ -49,7 +52,9 @@
     "suitcss-utils-display": "^1.0.2",
     "suitcss-utils-flex": "^2.0.0",
     "suitcss-utils-size": "^2.0.0",
-    "uglifyjs": "^2.4.11"
+    "uglifyjs": "^2.4.11",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11"
   },
   "devDependencies": {
     "autoprefixer": "^7.1.4",
@@ -77,11 +82,9 @@
     "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-lodash-fp": "^2.1.3",
     "eslint-plugin-react": "^7.4.0",
-    "extract-text-webpack-plugin": "^3.0.0",
     "flow-bin": "^0.56.0",
     "flux-standard-action": "^1.2.0",
     "gh-pages": "^1.0.0",
-    "html-webpack-plugin": "^2.28.0",
     "jest": "^21.2.1",
     "mockdate": "^2.0.1",
     "postcss-custom-media": "^6.0.0",
@@ -89,15 +92,13 @@
     "postcss-import": "^11.0.0",
     "postcss-loader": "^2.0.6",
     "prop-types": "^15.6.0",
-    "react-addons-test-utils": "^15.4.2",
-    "react-svg-loader": "^1.1.1",
+    "react-svg-loader": "^3.0.3",
     "react-test-renderer": "^16.0.0",
     "redux-mock-store": "^1.2.2",
     "style-loader": "^0.18.2",
     "uglify-js": "^3.1.3",
-    "uglifyjs-webpack-plugin": "^0.4.6",
-    "webpack": "^3.6.0",
-    "webpack-dev-server": "^2.4.0",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^4.0.0"
   }
 }
